+++ src/app/[locale]/products/page.tsx
@@ -6,7 +6,11 @@ const Products = async () => {
   const productsData = await getData("products");
 
   return (
-    <PageWrapper className="px-0" pageName="Products">
+    <PageWrapper
+      className="px-0"
+      pageName="Products"
+      dataForExport={productsData}
+    >
       <ProductsView products={productsData} />
     </PageWrapper>
   );