+++ src/components/common/Breadcrumbs.tsx
@@ -1,11 +1,15 @@
+import { useTranslations } from "next-intl";
+
 interface BreadcrumbsProps {
   pageName?: string;
 }
 
 export const Breadcrumbs = ({ pageName }: BreadcrumbsProps) => {
+  const t = useTranslations("breadcrumbs");
+
   return (
-    <div className="text-secondaryText text-1xl font-semibold">
-      Home &gt; {pageName}
+    <div className="text-secondaryText text-sm 1xl:text-base font-semibold">
+      {t("firstPart")} &gt; {t(pageName?.toLowerCase() as string)}
     </div>
   );
 };