+++ src/components/views/analytics/AssetPerformance.tsx
@@ -28,7 +28,7 @@ export const AssetPerformance = ({
       id="assetPerformance"
       title={t("title")}
     >
-      <Grid numItemsLg={3} className="mt-2 1xl:mt-4 3xl:mt-8 gap-x-10 gap-y-10">
+      <Grid numItemsLg={3} className="mt-4 1xl:mt-6 3xl:mt-8 gap-x-10 gap-y-10">
         <Flex>
           <DonutChart
             data={assetPerformanceData}